refactor: extract logic to check if non-mobile header is shrunken to hook
